Use `enum class` instead of `enum`, and autoformat the code (#652)
Enum class does less pollution of the namespace. This change involved some additional things like defining `&` operators since `enum class` isn't implicitly an int like `enum` is.

+++ src/stim/circuit/gate_data_table.h
@@ -34,10 +34,10 @@ struct GateVTable {
 #ifndef NDEBUG
         std::array<bool, NUM_DEFINED_GATES> seen{};
         for (const auto &[gate_id, value] : gate_data_pairs) {
-            seen[gate_id] = true;
+            seen[(size_t)gate_id] = true;
         }
         for (const auto &gate : GATE_DATA.items) {
-            if (!seen[gate.id]) {
+            if (!seen[(size_t)gate.id]) {
                 throw std::invalid_argument(
                     "Missing gate data! A value was not defined for '" + std::string(gate.name) + "'.");
             }
